Local Storage - Criando uma classe de erros personalizada

+++ src/storage/group/groupCreate.ts
@@ -1,4 +1,5 @@
 import AsyncStorage from '@react-native-async-storage/async-storage';
+import { AppError } from '@utils/AppError';
 
 import { groupsGetAll } from './groupsGetAll';
 import { GROUP_COLLECTION } from '@storage/storageConfig';
@@ -7,8 +8,15 @@ export async function groupCreate(newGroup: string) {
   try {
     const storedGroups = await groupsGetAll();
 
+    const groupAlreadyExists = storedGroups.includes(newGroup);
+
+    if(groupAlreadyExists) {
+      throw new AppError('Já existe um grupo cadastrado com esse nome.')
+    }
+
     const storage = JSON.stringify([...storedGroups, newGroup])
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
+    
   } catch (error) {
     throw error;
   }
